requirements.txt
streamlit
pytesseract
opencv-python-headless
pillow
numpy
